Fix Dark Mode Flicker and Handle Invalid Locales
Updated the dark mode logic to add a blocking js
script until the theme is set to avoid a brief
flicker of light mode on page load (when dark
mode is enabled).

Also updated the set locale plug to check the
given locale against the list of valid locales
and redirect to the home page if the locale
is not valid.

Finally, updated a helper function used by the
locale menu to redirect the user to the
expected locale when they click on one.